feat: MenuGroup 도메인
