+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -1,6 +1,7 @@
 package kitchenpos.menus.application;
 
 import kitchenpos.menus.domain.*;
+import kitchenpos.menus.tobe.domain.MenuGroup;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.infra.PurgomalumClient;
