+++ src/main/java/kitchenpos/menus/domain/MenuGroupRepository.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.domain;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.menus.tobe.domain.MenuGroup;
 
 public interface MenuGroupRepository {
     MenuGroup save(MenuGroup menuGroup);
