+++ src/main/java/kitchenpos/menus/tobe/domain/MenuGroup.java
@@ -1,0 +1,36 @@
+package kitchenpos.menus.tobe.domain;
+
+import javax.persistence.*;
+import java.util.Objects;
+import java.util.UUID;
+
+@Table(name = "menu_group")
+@Entity
+public class MenuGroup {
+    @Column(name = "id", columnDefinition = "binary(16)")
+    @Id
+    private UUID id;
+
+    @Column(name = "name", nullable = false)
+    @Embedded
+    private String name;
+
+    public MenuGroup() {
+    }
+
+    public MenuGroup(String name) {
+        if (Objects.isNull(name) || name.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        this.id = UUID.randomUUID();
+        this.name = name;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
